requirements.txt
streamlit
openai-whisper
librosa
numpy
tempfile
torch